api/logic changes
